Corrected Semantic Defintion of take(k)
I'm not sure whether I'm right here. But I assume, the deleted sentence is wrong here. 
If several consumers (threads) are waiting blocked by a take(k) for being satisfied and one of them wins when k is put in, then the other consumers should remain waiting, isn't it? Instead of returning null as the removed sentence is intending.
But as I said, I'm not sure what your initial intention was. In different context both may be meaningful.

May be it is worth to call a method covering you initial semantic (break blocking if another consumers wins) something like *takeIfWinning* or so.